+++ app/controllers/attendances_controller.rb
@@ -60,4 +60,6 @@ class AttendancesController < ApplicationController
   def attendance_params
     params.require(:attendance).permit(:search_param)
   end
+
+  def authorize!; end
 end